ENH Move _string_to_bool from axes._base to cbook
