+++ lib/matplotlib/axes/__init__.py
@@ -1,3 +1,2 @@
 from matplotlib.axes._subplots import *
 from matplotlib.axes._axes import *
-from matplotlib.axes._base import _string_to_bool
