+++ lib/mpl_toolkits/mplot3d/axes3d.py
@@ -1253,7 +1253,7 @@ class Axes3D(Axes):
         # TODO: Operate on each axes separately
         if len(kwargs) :
             b = True
-        self._draw_grid = maxes._string_to_bool(b)
+        self._draw_grid = cbook._string_to_bool(b)
 
     def ticklabel_format(self, **kwargs) :
         """
